# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Documentation for Furniture App API",
        "description": "Документация API для приложения Furniture App",
        "contact": {
            "email": "support@furnitureapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1/",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/colors": {
            "get": {
                "tags": [
                    "Цвета"
                ],
                "summary": "Получить все цвета",
                "description": "Возвращает список всех цветов.",
                "operationId": "9af5ded8d7143a0f97e53725eda97464",
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Color"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Цвета"
                ],
                "summary": "Создать новый цвет",
                "description": "Создаёт новый цвет и возвращает его.",
                "operationId": "e4d3770c69f36a71c5936b3b35655342",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Color"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Цвет создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Color"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ввод"
                    }
                }
            }
        },
        "/api/colors/{id}": {
            "get": {
                "tags": [
                    "Цвета"
                ],
                "summary": "Получить цвет по ID",
                "description": "Возвращает цвет по его ID.",
                "operationId": "c0e5a39d91e5093e2e0f77823615db5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Color"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Цвет не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Цвета"
                ],
                "summary": "Обновить цвет",
                "description": "Обновляет существующий цвет по его ID.",
                "operationId": "717be7234f38d1fb4ff4678ba4f59581",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Color"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Цвет обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Color"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Цвет не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Цвета"
                ],
                "summary": "Удалить цвет",
                "description": "Удаляет существующий цвет по его ID.",
                "operationId": "1667d6964960e3a8158e80e50c69f19c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Цвет удалён"
                    },
                    "404": {
                        "description": "Цвет не найден"
                    }
                }
            }
        },
...
        "/api/galleries": {
            "get": {
                "tags": [
                    "Галереи"
                ],
                "summary": "Получить все галереи",
                "description": "Возвращает список всех галерей.",
                "operationId": "4308cb10fec6f2c0e72356e3f9c2ea57",
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Gallery"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Галереи"
                ],
                "summary": "Создать новую галерею",
                "description": "Создаёт новую галерею и возвращает её.",
                "operationId": "e879c41d931b6fb537c5c19ff8eff892",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Gallery"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Галерея создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gallery"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ввод"
                    }
                }
            }
        },
        "/api/galleries/{id}": {
            "get": {
                "tags": [
                    "Галереи"
                ],
                "summary": "Получить галерею по ID",
                "description": "Возвращает галерею по её ID.",
                "operationId": "d55ab746a0aab8fa7e6fedfabece77fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gallery"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Галерея не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Галереи"
                ],
                "summary": "Обновить галерею",
                "description": "Обновляет существующую галерею по её ID.",
                "operationId": "b57576127d9ba5f3903b086f48fd3970",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Gallery"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Галерея обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Gallery"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Галерея не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Галереи"
                ],
                "summary": "Удалить галерею",
                "description": "Удаляет существующую галерею по её ID.",
                "operationId": "0789f145f187b6dad6df94da640d956c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Галерея удалена"
                    },
                    "404": {
                        "description": "Галерея не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Color": {
                "title": "Color",
                "description": "Color model",
                "required": [
                    "id",
                    "name",
                    "hex_code"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the color",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the color",
                        "type": "string"
                    },
                    "hex_code": {
                        "description": "Hexadecimal color code",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Gallery": {
                "title": "Gallery",
                "description": "Gallery model",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the gallery",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the gallery",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the gallery",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}